refactor path ops target setter
